improve resolution detection, inc. for retina displays
Retina displays aren't currently accounted for. Also there are conditions where a too low of resolution is auto-selected (for example "mobile" is selected instead of 720p if the resolution is higher than mobile but lower than 720p), which results in an image that is badly scaled.

Instead of directly comparing resolution to screen width, this is calculating the image density based on the devicePixelRatio and image resolution, and then selects the lowest possible image resolution that is above a certain image density value (0.7). Image density basically becomes an objective quality value that remains consistent when resolution and display density changes.